+++ ts/nni_manager/extensions/experiments_manager/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export { NNIExperimentsManager } from './manager';
